requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
keras
python-binance
